Separated FormatString from common.hpp for embedded use
